fixed float problem

+++ game.cpp
@@ -11,12 +11,13 @@ static void DrawTexFull(Texture2D tex, Rectangle dst)
 //Baum
 static void DrawTree(float x, float y, float s)
 {
-    DrawRectangle((int)(x - 5 * s), (int)(y - 22 * s), (int)(10 * s), (int)(22 * s),{101, 67, 33, 255});
-    DrawCircle((int)x, (int)(y - 30 * s), (int)(22 * s), {20, 100, 20, 255});
-    DrawCircle((int)(x - 10 * s), (int)(y - 42 * s), (int)(16 * s), {30, 130, 30, 255});
-    DrawCircle((int)(x + 10 * s), (int)(y - 42 * s), (int)(16 * s), {30, 130, 30, 255});
-    DrawCircle((int)x, (int)(y - 52 * s), (int)(18 * s), {50, 160, 50, 255});
+    DrawRectangle((int)(x - 5*s), (int)(y - 22*s), (int)(10*s), (int)(22*s), {101, 67, 33, 255});
+    DrawCircle((int)x,             (int)(y - 30*s), 22.0f*s, {20, 100, 20, 255});
+    DrawCircle((int)(x - 10*s),    (int)(y - 42*s), 16.0f*s, {30, 130, 30, 255});
+    DrawCircle((int)(x + 10*s),    (int)(y - 42*s), 16.0f*s, {30, 130, 30, 255});
+    DrawCircle((int)x,             (int)(y - 52*s), 18.0f*s, {50, 160, 50, 255});
 }
+
 //Kaktus
 static void DrawCactus(float x, float y, float s)
 {
@@ -31,29 +32,31 @@ static void DrawCactus(float x, float y, float s)
 
 static void DrawUFO(float x, float y, float s)
 {
-    DrawEllipse((int)x, (int)(y + 4 * s), (int)(26 * s), (int)(7 * s), {80, 80, 100, 255});
-    DrawCircle((int)(x - 12 * s), (int)(y + 7 * s), (int)(4 * s), {255, 140, 0, 200});
-    DrawCircle((int)x, (int)(y + 9 * s), (int)(5 * s), {255, 100, 0, 200});
-    DrawCircle((int)(x + 12 * s), (int)(y + 7 * s), (int)(4 * s), {255, 140, 0, 200});
-    DrawEllipse((int)x, (int)y, (int)(28 * s), (int)(9 * s), {160, 160, 185, 255});
-    DrawEllipseLines((int)x, (int)y, (int)(28 * s), (int)(9 * s), {220, 220, 255, 180});
-    DrawEllipse((int)x, (int)(y - 7 * s), (int)(13 * s), (int)(10 * s), {80, 200, 255, 210});
-    DrawEllipseLines((int)x, (int)(y - 7 * s), (int)(13 * s), (int)(10 * s), {180, 240, 255, 255});
-    DrawCircle((int)(x - 18 * s), (int)(y + 2 * s), (int)(3 * s), {0, 255, 120, 220});
-    DrawCircle((int)(x + 18 * s), (int)(y + 2 * s), (int)(3 * s), {255, 50, 50, 220});
+    DrawEllipse((int)x, (int)(y + 4*s), 26.0f*s, 7.0f*s, {80, 80, 100, 255});
+    DrawCircle((int)(x - 12*s), (int)(y + 7*s), 4.0f*s, {255, 140, 0, 200});
+    DrawCircle((int)x,          (int)(y + 9*s), 5.0f*s, {255, 100, 0, 200});
+    DrawCircle((int)(x + 12*s), (int)(y + 7*s), 4.0f*s, {255, 140, 0, 200});
+    DrawEllipse((int)x, (int)y, 28.0f*s, 9.0f*s, {160, 160, 185, 255});
+    DrawEllipseLines((int)x, (int)y, 28.0f*s, 9.0f*s, {220, 220, 255, 180});
+    DrawEllipse((int)x, (int)(y - 7*s), 13.0f*s, 10.0f*s, {80, 200, 255, 210});
+    DrawEllipseLines((int)x, (int)(y - 7*s), 13.0f*s, 10.0f*s, {180, 240, 255, 255});
+    DrawCircle((int)(x - 18*s), (int)(y + 2*s), 3.0f*s, {0, 255, 120, 220});
+    DrawCircle((int)(x + 18*s), (int)(y + 2*s), 3.0f*s, {255, 50, 50, 220});
 }
 
 static void DrawPlanet(float x, float y, float s, int variant)
 {
-    Color colors[] = {{220, 80, 80, 255}, {80, 160, 255, 255}, {255, 180, 60, 255}, {180, 80, 255, 255}};
+    Color colors[] = {{220, 80,  80,  255}, {80,  160, 255, 255},
+                      {255, 180, 60,  255}, {180, 80,  255, 255}};
     Color c = colors[variant % 4];
-    int r = (int)(14 * s);
+    float r = 14.0f * s;
     DrawCircle((int)x, (int)y, r, c);
-    DrawCircle((int)(x + r * 0.25f), (int)(y - r * 0.25f), (int)(r * 0.55f), Fade(WHITE, 0.15f));
+    DrawCircle((int)(x + r*0.25f), (int)(y - r*0.25f), r * 0.55f, Fade(WHITE, 0.15f));
     DrawCircleLines((int)x, (int)y, r, Fade(WHITE, 0.3f));
     if (variant % 2 == 1)
-        DrawEllipseLines((int)x, (int)y, (int)(22 * s), (int)(5 * s), Fade(WHITE, 0.5f));
+        DrawEllipseLines((int)x, (int)y, 22.0f*s, 5.0f*s, Fade(WHITE, 0.5f));
 }
+
 
 // --- Init ---
 void Game::Init()
